разбила код на модули, добавлены три файла: description.js; database.js; getRandom.js
