+++ js/database.js
@@ -1,0 +1,56 @@
+const TITLE_LIST = [
+  'Название одно',
+  'Название второе',
+  'Название третье',
+];
+
+const CHECK_IN = [
+  '12:00',
+  '13:00',
+  '14:00'
+];
+
+const CHECK_OUT = [
+  '12:00',
+  '13:00',
+  '14:00'
+];
+
+const TYPE = [
+  'palace',
+  'flat',
+  'house',
+  'bungalow',
+  'hotel'
+];
+
+const FEATURES = [
+  'wifi',
+  'dishwasher',
+  'parking',
+  'washer',
+  'elevator',
+  'conditioner'
+];
+
+const DESCRIPTION = [
+  'Описание один',
+  'Описание два',
+  'Описание три'
+];
+
+const PHOTOS = [
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/duonguyen-8LrGtIxxa4w.jpg',
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/brandon-hoogenboom-SNxQGWxZQi0.jpg',
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/claire-rendall-b6kAwr1i0Iw.jpg'
+];
+
+export {
+  TITLE_LIST,
+  CHECK_IN,
+  CHECK_OUT,
+  TYPE,
+  FEATURES,
+  DESCRIPTION,
+  PHOTOS
+};
